Moved TokenCreator back into MainVerticle
This allows the failStartOnInvalidAlgorithm test to pass. Also fixed up that
test to not match on a string and not use a depreciated form of assertThat.
I think this is fine since I am imagining that other new apis will need/want to
create tokens. We may also want to move other shared resource creation out into
MainVerticle like the perms and user services.
