+++ src/test/java/org/folio/auth/authtokenmodule/MainVerticleTest.java
@@ -1,9 +1,5 @@
 package org.folio.auth.authtokenmodule;
 
-import static org.hamcrest.CoreMatchers.containsString;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertThrows;
-
 import com.nimbusds.jose.JOSEException;
 
 import org.apache.logging.log4j.Level;
@@ -14,6 +10,10 @@ import io.vertx.core.Vertx;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.hamcrest.MatcherAssert.assertThat;
+
 @RunWith(VertxUnitRunner.class)
 public class MainVerticleTest {
   @Test
@@ -23,34 +23,20 @@ public class MainVerticleTest {
         context.asyncAssertSuccess(run -> vertx.close(context.asyncAssertSuccess())));
   }
 
-  // @Test
-  // public void failStartOnInvalidAlgorithm(TestContext context) {
-  //   // class MainVerticleInvalidAlgorithm extends AuthorizeApi {
-  //   //   @Override
-  //   //   TokenCreator getTokenCreator() throws JOSEException {
-  //   //     TokenCreator tokenCreator = super.getTokenCreator();
-  //   //     tokenCreator.setJweHeader(null);
-  //   //     return tokenCreator;
-  //   //   }
-  //   // }
+  @Test
+  public void failStartOnInvalidAlgorithm(TestContext context) {
+    class MainVerticleInvalidAlgorithm extends MainVerticle {
+      @Override
+      TokenCreator getTokenCreator() throws JOSEException {
+        TokenCreator tokenCreator = super.getTokenCreator();
+        tokenCreator.setJweHeader(null);
+        return tokenCreator;
+      }
+    }
 
-  //   // Vertx.vertx().deployVerticle(new MainVerticleInvalidAlgorithm(null, null),
-  //   //     context.asyncAssertFailure(fail -> assertThat(fail.getMessage(), containsString("TokenCreator"))));
-  //   class MainVerticleInvalidAlgorithm extends AuthorizeApi {
-  //     @Override
-  //     TokenCreator getTokenCreator() throws JOSEException {
-  //       TokenCreator tokenCreator = super.getTokenCreator();
-  //       tokenCreator.setJweHeader(null);
-  //       return tokenCreator;
-  //     }
-  //     public MainVerticleInvalidAlgorithm(Vertx vertx) {}
-  //   }
-
-  //   // Vertx.vertx().deployVerticle(new MainVerticleInvalidAlgorithm(null, null),
-  //   //     context.asyncAssertFailure(fail -> assertThat(fail.getMessage(), containsString("TokenCreator"))));
-  //   assertThrows(MissingAlgorithmException.class, () -> new MainVerticleInvalidAlgorithm(null));
-  //   //assertThrows(RuntimeException.class, () -> new AuthorizeApi(null));
-  // }
+    Vertx.vertx().deployVerticle(new MainVerticleInvalidAlgorithm(),
+         context.asyncAssertFailure(fail -> assertThat(fail, is(instanceOf(MissingAlgorithmException.class)))));
+  }
 
   @Test
   public void setLogLevel(TestContext context) {
